+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "dev": "nodemon --exec babel-node ./server/server.js",
-    "start": "babel-node ./server/server.js",
+    "dev": "nodemon --exec babel-node ./server/app.js",
+    "start": "babel-node ./server/app.js",
     "test": "nyc --reporter=text --reporter=html mocha server/test/allTest.js --require @babel/polyfill  --require @babel/register --exit",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
