--- server/server.js
@@ -1,7 +1,0 @@
-import app from './app';
-
-const PORT = process.env.PORT || 5000;
-
-const server = app.listen(PORT, () => console.log(`Server is running on port: ${PORT}`));
-
-export default server;
